requirements.txt
# Dependencies
streamlit==1.28.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
nltk==3.8.1
spacy==3.6.1
transformers==4.34.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
networkx==3.1
rouge-score==0.1.2
PyPDF2==3.0.1
python-docx==0.8.11
sqlite3
pickle-mixin==1.0.2
textstat==0.7.3
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
